添加随机获取题目的接口，修改所有主键类型为Long

+++ db/quiz.sql
@@ -1,6 +1,6 @@
 -- 用户表
 CREATE TABLE user (
-    id INT AUTO_INCREMENT PRIMARY KEY,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
     account VARCHAR(32) UNIQUE COMMENT '账号',
     name VARCHAR(32) NOT NULL COMMENT '用户名',
     password VARCHAR(255) NOT NULL COMMENT '密码',
@@ -11,18 +11,18 @@ CREATE TABLE user (
 
 -- 题库表
 CREATE TABLE question_bank (
-    id INT AUTO_INCREMENT PRIMARY KEY,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(100) NOT NULL COMMENT '题库名',
     subject VARCHAR(100) NOT NULL COMMENT '学科分类（如数学、编程）',
-    creator_id INT COMMENT '创建者ID',
+    creator_id BIGINT COMMENT '创建者ID',
     create_time DATETIME DEFAULT CURRENT_TIMESTAMP COMMENT '题库创建时间',
     FOREIGN KEY (creator_id) REFERENCES user (id)
 ) COMMENT = '题库表';
 
 -- 题目表
 CREATE TABLE question (
-    id INT AUTO_INCREMENT PRIMARY KEY,
-    bank_id INT COMMENT '所属题库ID',
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    bank_id BIGINT COMMENT '所属题库ID',
     type ENUM(
         'SINGLE_CHOICE',
         'MULTIPLE_CHOICE',
@@ -33,7 +33,7 @@ CREATE TABLE question (
     answer JSON NOT NULL COMMENT '题目选项和答案 {options,answer}/{answer}',
     explanation TEXT COMMENT '题目解析',
     difficulty ENUM('EASY', 'MEDIUM', 'HARD') NOT NULL COMMENT '题目难度',
-    creator_id INT COMMENT '创建者ID',
+    creator_id BIGINT COMMENT '创建者ID',
     create_time DATETIME DEFAULT CURRENT_TIMESTAMP COMMENT '题目创建时间',
     FOREIGN KEY (bank_id) REFERENCES question_bank (id),
     FOREIGN KEY (creator_id) REFERENCES user (id)
@@ -41,21 +41,21 @@ CREATE TABLE question (
 
 -- 试卷表
 CREATE TABLE paper (
-    id INT AUTO_INCREMENT PRIMARY KEY,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(100) NOT NULL COMMENT '试卷名称',
     time_limit INT COMMENT '考试时间限制(分钟,0表示不限时)',
     difficulty ENUM('EASY', 'MEDIUM', 'HARD') NOT NULL COMMENT '试卷难度',
     total_score INT NOT NULL DEFAULT 0 COMMENT '试卷总分',
-    creator_id INT COMMENT '创建者ID',
+    creator_id BIGINT COMMENT '创建者ID',
     create_time DATETIME DEFAULT CURRENT_TIMESTAMP COMMENT '试卷创建时间',
     FOREIGN KEY (creator_id) REFERENCES user (id)
 ) COMMENT = '试卷表';
 
 -- 试卷题目关联表
 CREATE TABLE paper_question (
-    id INT AUTO_INCREMENT PRIMARY KEY,
-    paper_id INT COMMENT '试卷ID',
-    question_id INT COMMENT '题目ID',
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    paper_id BIGINT COMMENT '试卷ID',
+    question_id BIGINT COMMENT '题目ID',
     score INT COMMENT '题目分数',
     FOREIGN KEY (paper_id) REFERENCES paper (id),
     FOREIGN KEY (question_id) REFERENCES question (id),
@@ -64,9 +64,9 @@ CREATE TABLE paper_question (
 
 -- 考试记录表
 CREATE TABLE exam (
-    id INT AUTO_INCREMENT PRIMARY KEY,
-    user_id INT NOT NULL COMMENT '用户ID',
-    paper_id INT NOT NULL COMMENT '试卷ID',
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    user_id BIGINT NOT NULL COMMENT '用户ID',
+    paper_id BIGINT NOT NULL COMMENT '试卷ID',
     score INT NOT NULL COMMENT '得分',
     start_time DATETIME COMMENT '开始答题时间',
     end_time DATETIME COMMENT '结束答题时间',
@@ -76,10 +76,10 @@ CREATE TABLE exam (
 
 -- 答题记录表
 CREATE TABLE ans_record (
-    id INT AUTO_INCREMENT PRIMARY KEY,
-    user_id INT NOT NULL COMMENT '用户ID',
-    question_id INT NOT NULL COMMENT '题目ID',
-    exam_id INT COMMENT '考试记录ID',
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    user_id BIGINT NOT NULL COMMENT '用户ID',
+    question_id BIGINT NOT NULL COMMENT '题目ID',
+    exam_id BIGINT COMMENT '考试记录ID',
     user_answer VARCHAR(255) NOT NULL COMMENT '用户提交的答案',
     start_time DATETIME COMMENT '开始答题时间',
     end_time DATETIME COMMENT '结束答题时间',
@@ -90,9 +90,9 @@ CREATE TABLE ans_record (
 
 -- 错题记录表
 CREATE TABLE mistake (
-    id INT AUTO_INCREMENT PRIMARY KEY,
-    user_id INT NOT NULL COMMENT '用户ID',
-    question_id INT NOT NULL COMMENT '题目ID',
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    user_id BIGINT NOT NULL COMMENT '用户ID',
+    question_id BIGINT NOT NULL COMMENT '题目ID',
     user_answer VARCHAR(255) NOT NULL COMMENT '用户提交的答案',
     FOREIGN KEY (user_id) REFERENCES user (id),
     FOREIGN KEY (question_id) REFERENCES question (id)
